🗺️ Atlas: Splitting Config God Module
🕸️ Tangle: `src/config.rs` was over 1600 lines long, a "God Module" containing configurations for WAL, Historical Storage, Vector Index, and the overall AletheiaDB config. This violated the Single Responsibility Principle and made navigation difficult.
📐 Blueprint: Refactored `src/config.rs` into a `src/config/` directory. Extracted configurations into cohesive submodules: `wal.rs`, `historical.rs`, `vector.rs`, `db.rs`, `error.rs`, and `env.rs`. Extracted tests to `tests.rs`. Kept `mod.rs` as a clean facade, using `pub use` to maintain the exact public API, preventing any upstream breakage.
🧱 Stability: Improved navigability and cohesion without breaking public API.
🔬 Verification: `cargo test config::` passes successfully.
